Making line.js
PATH IS LINE.
if I am not painting(painting is false), just create "path" and move the path. Move the starting point of a new sub-path to the (x,y) coordinates.
and if i am painting(painting is true),
make this path above to line. Path becomes stroke line with lineTo(x,y) methods.